requirements.txt
numpy
pandas
scikit-learn
tensorflow
keras
matplotlib
seaborn
streamlit
pillow
plotly
opencv-python
kaggle
xgboost
seaborn
inference-sdk
transformers
huggingface_hub
torch